hoist all the market related data fetch logic to global context to avoid duplicated requests, thus improve performance
